webapp/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-bootstrap4==22.3
idna==3.4
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.13
requests
pendulum
Pillow
icmplib
Django
django-bootstrap4==22.3
mysql-connector-python
django-crispy-forms==1.14.0

sqlalchemy>=1.2.18
celery>=5.0.5
django-celery-results
Redis
flower
mysqlclient
django-celery-beat

python-nmap


#Vrec
zeep
pandas
openpyxl
suds

pydig

#ssh
paramiko